добавлены свайпы в избранном и удаление рецепта и кроссРефов по свайпу. + вибрация

+++ app/src/main/java/ru/geekbrains/bookofrecipes/data/RecipesRepository.kt
@@ -46,7 +46,6 @@ class RecipesRepository(
             localDataSource.getRecipeList()
 
 
-    override suspend fun deleteRecipeFromFavorites(recipe: RecipeInformation): Either<Failure, Int> =
+    override suspend fun deleteRecipeFromFavorites(recipe: RecipeInformation): Either<Failure, Long> =
         localDataSource.deleteRecipe(recipe)
-
 }
