+++ app/src/main/java/ru/geekbrains/bookofrecipes/data/local/LocalDataSource.kt
@@ -9,5 +9,5 @@ interface LocalDataSource {
 
     suspend fun saveRecipe(recipe: RecipeInformation): Either<Failure, Long>
 
-    suspend fun deleteRecipe(recipe: RecipeInformation): Either<Failure, Int>
+    suspend fun deleteRecipe(recipe: RecipeInformation): Either<Failure, Long>
 }
